OAK-1645 - Route find queries to Mongo secondary in MongoDocumentStore
The find and query method now explicitly define the read preference to be used. The preference is determined by type of Collection , maxCacheAge and modified time of parent. if preference is secondary then actual preference used is picked from DB level settings

Also exposed a setting 'maxReplicationLagMillis' which determines the max possible lag in replication beyond which its safe to route the read queries to secondaries