OAK-3591 - Lucene index with 'analyzed=true' sometimes used by mistake
PropertyDefinition
If index=false then all other config are assumed to be false ignoring whatever is configured. As if property is not be indexed then it is not considered for all other cases also

IndexDefinition
Introduced a new `nodeFullTextIndexed` field to treat cases where node is indexed from cases where only certain property are index with analyzed=true. This ensures that `areAlMatchingNodeByTypeIndexed` would only return true if there is a nodeScope indexed field. So far it was returning true even when there was only a analyzed field

IndexPlanner
For a fulltext query now if the fulltext term is based on node then it only answers true if and only if node is indexed for fulltext i.e. there at least one property config with nodeScopeIndex=true

git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1714520 13f79535-47bb-0310-9956-ffa450edef68