OAK-4412 - Lucene hybrid index
Synchronous indexing support continue - Simplified the refresh policy to take in a callback which is responsible for refreshing the readers if the index is found to be updated.

For sync indexes the refresh is done by the writer in DocumentQueue i.e. its done as part of commit thread itself.