OAK-4939 - Isolate corrupted index and make async indexer more resilient
Introducing a TrackingCorruptIndexHandler which would track failing indexes. If an index is failing for long time (default 30 mins) then it would be marked as corrupt.

The index editors must call IndexingContext#indexUpdateFailed on passed callback instance to indicate that index update has failed and that index should be tracked for failure

The setting of corrupt flag is done as part of IndexUpdate commit itself as IndexUpdate reads the value of 'corrupt' flag from NodeBuilder. Doing this separately would require reading from different state hence avoided